requirements.txt
Django==1.9
django-material==0.7.0
django-registration-redux==1.4
psycopg2==2.6.2